# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "校园活动管理系统 API",
        "description": "校园活动管理系统接口文档",
        "contact": {
            "name": "开发团队",
            "email": "example@xidian.edu.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
...
    "paths": {
...
        "/registration/checkin": {
            "post": {
                "tags": [
                    "报名管理"
                ],
                "summary": "学生签到",
                "description": "学生签到参加活动",
                "operationId": "checkIn",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckInDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultRegistrationVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/registration/checkin-by-token": {
            "post": {
                "tags": [
                    "报名管理"
                ],
                "summary": "扫码签到",
                "description": "通过扫描二维码获取Token进行签到",
                "operationId": "checkInByToken",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckInByTokenDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultRegistrationVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/h5/checkin": {
            "post": {
                "tags": [
                    "H5移动端接口"
                ],
                "summary": "H5扫码签到",
                "description": "通过Token和手机号完成签到",
                "operationId": "checkIn_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckInByTokenDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultRegistrationVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/h5/config": {
            "get": {
                "tags": [
                    "H5移动端接口"
                ],
                "summary": "获取H5配置",
                "description": "获取H5页面配置信息",
                "operationId": "getPageConfig",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultPageConfig"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/h5/checkin/validate": {
            "get": {
                "tags": [
                    "H5移动端接口"
                ],
                "summary": "验证签到Token",
                "description": "扫描二维码后验证Token是否有效",
                "operationId": "validateCheckInToken",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "签到Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ResultVoid": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
...
            "RegistrationVO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "报名ID",
                        "format": "int64",
                        "example": 1
                    },
                    "activityId": {
                        "type": "integer",
                        "description": "活动ID",
                        "format": "int64",
                        "example": 1
                    },
                    "activityName": {
                        "type": "string",
                        "description": "活动名称",
                        "example": "新生计算机基础讲座"
                    },
                    "studentName": {
                        "type": "string",
                        "description": "学生姓名",
                        "example": "张三"
                    },
                    "studentPhone": {
                        "type": "string",
                        "description": "学生手机号",
                        "example": "13800138000"
                    },
                    "studentCollege": {
                        "type": "string",
                        "description": "学生学院",
                        "example": "计算机学院"
                    },
                    "registrationStatus": {
                        "type": "integer",
                        "description": "报名状态：1-报名成功, 2-已取消",
                        "format": "int32",
                        "example": 1
                    },
                    "registrationStatusName": {
                        "type": "string",
                        "description": "报名状态名称",
                        "example": "报名成功"
                    },
                    "checkInStatus": {
                        "type": "integer",
                        "description": "签到状态：0-未签到, 1-已签到",
                        "format": "int32",
                        "example": 0
                    },
                    "checkInStatusName": {
                        "type": "string",
                        "description": "签到状态名称",
                        "example": "未签到"
                    },
                    "checkInTime": {
                        "type": "string",
                        "description": "签到时间",
                        "format": "date-time"
                    },
                    "remarks": {
                        "type": "string",
                        "description": "备注信息",
                        "example": "希望参加此活动学习相关知识"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "报名时间",
                        "format": "date-time"
                    },
                    "updateTime": {
                        "type": "string",
                        "description": "更新时间",
                        "format": "date-time"
                    },
                    "canCancel": {
                        "type": "boolean",
                        "description": "是否可以取消报名",
                        "example": true
                    },
                    "canCheckIn": {
                        "type": "boolean",
                        "description": "是否可以签到",
                        "example": true
                    }
                },
                "description": "报名信息"
            },
            "ResultRegistrationVO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/RegistrationVO"
                    }
                }
            },
            "CheckInDTO": {
                "required": [
                    "activityId",
                    "studentPhone"
                ],
                "type": "object",
                "properties": {
                    "activityId": {
                        "type": "integer",
                        "description": "活动ID",
                        "format": "int64",
                        "example": 1
                    },
                    "studentPhone": {
                        "pattern": "^1[3-9]\\d{9}$",
                        "type": "string",
                        "description": "学生手机号",
                        "example": "13800138000"
                    }
                },
                "description": "签到信息"
            },
            "CheckInByTokenDTO": {
                "required": [
                    "checkInToken",
                    "studentPhone"
                ],
                "type": "object",
                "properties": {
                    "checkInToken": {
                        "type": "string",
                        "description": "签到Token(从二维码获取)",
                        "example": "eyJhbGciOiJIUzI1NiJ9..."
                    },
                    "studentPhone": {
                        "pattern": "^1[3-9]\\d{9}$",
                        "type": "string",
                        "description": "学生手机号",
                        "example": "13800138000"
                    }
                },
                "description": "签到信息"
            },
...
            "PageConfig": {
                "type": "object",
                "properties": {
                    "appName": {
                        "type": "string"
                    },
                    "checkInTitle": {
                        "type": "string"
                    },
                    "phonePlaceholder": {
                        "type": "string"
                    },
                    "submitButtonText": {
                        "type": "string"
                    }
                }
            },
            "ResultPageConfig": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageConfig"
                    }
                }
            },
            "ResultMapStringObject": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer Authentication": {
                "type": "http",
                "description": "请输入JWT Token，格式：直接输入token即可（不需要Bearer前缀）",
                "name": "Bearer Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}